Time tracking now uses long (>18h) and use struct for conversions.

+++ src/main.h
@@ -14,7 +14,7 @@
 
 typedef struct hbconfig hbconfig;
 struct hbconfig {
-  volatile unsigned int time;
+  volatile unsigned long time;
   volatile unsigned char temp;
 };
